--- ui/src/components/Configs/index.tsx
@@ -1,3 +1,0 @@
-export * from './StartConfigs';
-export * from './UpsertConfig';
-export * from './OnBoarding';
