+++ ui/src/components/Feedback/DownloadProgress/DownloadProgress.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, useEffect, useState } from 'react';
-import { useDDClient } from '../../services/hooks';
+import { useDDClient } from '../../../services/hooks';
 import { CircularProgressWithLabel } from './CircularProgressWithLabel';
