+++ ui/src/components/Feedback/DownloadProgress/index.tsx
@@ -1,0 +1,2 @@
+export * from './DownloadProgress';
+export * from './CircularProgressWithLabel';
