+++ ui/src/components/Feedback/index.tsx
@@ -1,0 +1,2 @@
+export * from './ConfirmableButton';
+export * from './DownloadProgress/index';
