+++ ui/src/components/Header/Menu.tsx
@@ -3,8 +3,8 @@ import { IconButton, Menu } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import { useDDClient } from '../../services/hooks';
 import { DockerImage } from '../../types';
-import { ConfirmableButton } from '../Configs/ConfirmableButton';
-import { UpdateDialog } from '../Dialog/UpdateDialog';
+import { ConfirmableButton } from '../Feedback/ConfirmableButton';
+import { UpdateDialog } from '../Views';
 
 const ITEM_HEIGHT = 80;
