+++ ui/src/components/Views/SystemStatus/StatusPage.tsx
@@ -6,11 +6,11 @@ import {
 } from '@mui/material';
 import { createStyles, makeStyles } from '@mui/styles';
 import { Refresh } from '@mui/icons-material';
-import { useLocalStackHealth } from '../../services/hooks/health';
-import { HealthState } from '../../types';
-import { Capitalize } from '../../services/generic/utils';
-import { Status as SystemStatusIcon } from './Status';
-import { useLocalStack } from '../../services/hooks';
+import { useLocalStackHealth } from '../../../services/hooks/health';
+import { HealthState } from '../../../types';
+import { Capitalize } from '../../../services/generic/utils';
+import { SystemStatus as SystemStatusIcon } from './SystemStatus';
+import { useLocalStack } from '../../../services/hooks';
 
 const ORDER = [
   HealthState.RUNNING,
@@ -37,7 +37,7 @@ const useStyles = makeStyles((theme: Theme) => createStyles({
   },
 }));
 
-export const SystemStatus = (): ReactElement => {
+export const StatusPage = (): ReactElement => {
   const { health, mutate } = useLocalStackHealth();
   const { data } = useLocalStack();
