+++ ui/src/components/Views/SystemStatus/index.tsx
@@ -1,2 +1,2 @@
-export * from './Status';
 export * from './SystemStatus';
+export * from './StatusPage';
