+++ ui/src/components/Views/index.tsx
@@ -1,0 +1,4 @@
+export * from './Configs';
+export * from './SystemStatus';
+export * from './Logs/LogsPage';
+export * from './Update/UpdateDialog';
