+++ ui/src/components/index.ts
@@ -1,6 +1,4 @@
-export * from './Configs';
+export * from './Views/Configs';
 export * from './TabPanel/ControlledTabPanel';
 export * from './Header';
-export * from './Logs/Logs';
-export * from './SystemStatus';
-export * from './Dialog/UpdateDialog';
+export * from './Views';
